requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-filter==2.3.0
django-phonenumber-field==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
idna==2.10
mailchimp-marketing==3.0.15
Markdown==3.2.2
phonenumbers==8.12.9
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
